Relocation of the new service files
