+++ tools/gestenrekorder/src/app/touch-area/service/circle-renderer.service.spec.ts
