+++ tools/gestenrekorder/src/app/touch-area/service/event.service.spec.ts
